# Conflicts:
#	GameTest/GameTest.vcxproj
#	GameTest/GameTest/Scenes/MainScene.cpp
#	PM3DApi/Api/Private/Camera/Camera.cpp
#	PM3DApi/Api/Public/Camera/Camera.h
#	PM3DApi/Api/Public/EventSystem/EventSystem.h
#	PM3DApi/Api/Public/Util/Util.h
#	PM3DApi/PM3DApi.vcxproj

+++ GameTest/GameTest/Components/LightMoverComponent.h
@@ -7,7 +7,7 @@
 class LightMoverComponent : public PM3D_API::Component
 {
 public:
-    explicit LightMoverComponent(const int offset) : angle(offset)
+    explicit LightMoverComponent(const float offset) : angle(offset)
     {
         
     }
